[refactor] Context and useHealth hook
