+++ renderer/pages/index.tsx
@@ -8,13 +8,8 @@ import PlayBtn from '../components/buttons/PlayBtn'
 import ShopBtn from '../components/buttons/ShopBtn'
 import StartBtn from '../components/buttons/StartBtn'
 import MiniCat from '../components/characters/MiniCat'
-import { HealthContext } from '../components/containers/CanvasContext'
+import { GameContext } from '../components/containers/CanvasContext'
 import BackGround from '../components/items/BackGround'
-import Coin from '../components/items/Coin'
-import CuteFish from '../components/items/CuteFish'
-import Level from '../components/items/Level'
-import LevelBar from '../components/items/LevelBar'
-import LifeGauge from '../components/items/LifeGauge'
 import NumText from '../components/items/NumText'
 import TitleBar from '../components/items/TitleBar'
 import Black from '../components/items/black'
@@ -34,7 +29,7 @@ import {
 
 const IndexPage = () => {
   const router = useRouter()
-  const { health } = useContext(HealthContext)
+  const { health } = useContext(GameContext)
 
   const [windowPosition, setWindowPosition] = useState<Position>({
     x: 350,
@@ -143,32 +138,6 @@ const IndexPage = () => {
         }}
       />
       <BackGround>
-        <Container x={1170} y={180} scale={1}>
-          <LevelBar n={ex.progress(8)} x={-60} y={-30} scale={0.75} />
-          <Level x={130} y={170} scale={0.7} />
-          <NumText
-            n={ex.get_level()}
-            view_digits={3}
-            x={80}
-            y={140}
-            scale={0.65}
-            is_headzero_displayed={true}
-          />
-        </Container>
-        <Container x={1150} y={600} scale={0.6}>
-          <Coin x={-20} scale={0.8} />
-          <NumText n={coins} view_digits={4} x={50} y={-25} scale={0.7} />
-        </Container>
-        <LifeGauge
-          n={hp.get_health_point_formatted(10)}
-          x={1200}
-          y={510}
-          scale={1.2}
-        />
-        <Container x={1300} y={600} scale={0.6}>
-          <CuteFish x={60} y={8} scale={0.4} />
-          <NumText n={fish} view_digits={4} x={150} y={-25} scale={0.7} />
-        </Container>
         <Sprite image="/static/img/board.png" x={40} scale={0.8} />
         <Sprite image="/static/img/days.png" x={300} y={190} scale={0.5} />
         {/*  一日目からスタート */}
